refactor: 변수명 변경 및 함수명 변경
